requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.8.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.7
cryptography==46.0.5
distro==1.9.0
faiss-cpu==1.10.0
Flask==3.0.3
google-auth==2.49.1
google-genai==1.67.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.2.2
packaging==26.0
pdf2image==1.17.0
pillow==12.1.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==5.2.0
requests==2.32.3
sniffio==1.3.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
Werkzeug==3.0.6
